+++ src/main/java/com/capstone/godofinterview/domain/auth/dto/response/LoginResponse.java
@@ -1,8 +1,10 @@
-package com.capstone.godofinterview.domain.auth.dto;
+package com.capstone.godofinterview.domain.auth.dto.response;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
 @AllArgsConstructor
+@Getter
 public class LoginResponse {
 
     private String accessToken;
